+++ stack/__init__.py
@@ -8,8 +8,11 @@ USE_ECS = os.environ.get("USE_ECS") == "on"
 USE_EKS = os.environ.get("USE_EKS") == "on"
 USE_GOVCLOUD = os.environ.get("USE_GOVCLOUD") == "on"
 USE_NAT_GATEWAY = os.environ.get("USE_NAT_GATEWAY") == "on"
+USE_CLOUDFRONT = os.environ.get("USE_CLOUDFRONT") == "on"
 
-if USE_EKS:
+if USE_CLOUDFRONT:
+    from . import cdn  # noqa: F401
+elif USE_EKS:
     from . import assets  # noqa: F401
     from . import cache  # noqa: F401
     from . import database  # noqa: F401
@@ -22,7 +25,6 @@ if USE_EKS:
     if not USE_GOVCLOUD:
         # make sure this isn't added to the template for GovCloud, as it's not
         # supported in this region
-        from . import cdn  # noqa: F401
         from . import search  # noqa: F401
 
     if USE_NAT_GATEWAY:
@@ -40,7 +42,6 @@ else:
     if not USE_GOVCLOUD:
         # make sure this isn't added to the template for GovCloud, as it's not
         # supported in this region
-        from . import cdn  # noqa: F401
         from . import search  # noqa: F401
 
     if USE_NAT_GATEWAY:
