+++ stack/environment.py
@@ -1,16 +1,7 @@
-import os
-
 from troposphere import AWS_REGION, GetAtt, If, Join, Ref
 
 from . import USE_GOVCLOUD
-from .assets import (
-    assets_bucket,
-    assets_cloudfront_domain,
-    assets_custom_domain_condition,
-    assets_use_cloudfront_condition,
-    distribution,
-    private_assets_bucket
-)
+from .assets import assets_bucket, private_assets_bucket
 from .cache import cache_url, redis_url
 from .common import secret_key
 from .database import (
@@ -32,15 +23,6 @@ if not USE_GOVCLOUD:
 else:
     es_domain = None
 
-
-if os.getenv('USE_GOVCLOUD') != 'on' and os.getenv('USE_EB') != 'on':
-    # GovCloud doesn't support CloudFront, and with Elastic Beanstalk
-    # attempting to add an environment variable pointing to to the distrubtion
-    # creates a circular dependency (because it's not possible to create the
-    # EB load balancer separately from the EB application itself)
-    from .cdn import app_distribution, app_uses_cloudfront_condition
-else:
-    app_distribution = None
 
 environment_variables = [
     ("AWS_REGION", Ref(AWS_REGION)),
@@ -86,39 +68,6 @@ environment_variables = [
     ("CACHE_URL", cache_url),
     ("REDIS_URL", redis_url),
 ]
-
-if distribution:
-    # not supported by GovCloud, so add it only if it was created
-    environment_variables.append(
-        ("CDN_DOMAIN_NAME", If(
-            assets_use_cloudfront_condition,
-            If(
-                # use the custom domain passed into the stack, otherwise fallback to the default domain
-                assets_custom_domain_condition,
-                Ref(assets_cloudfront_domain),
-                GetAtt(distribution, "DomainName"),
-            ),
-            "",
-        )),
-    )
-
-if app_distribution and os.getenv('USE_EB') != 'on':
-    # Not supported by GovCloud, so add it only if it was created. Also, EB
-    # attempting to add an environment variable pointing to the distrubtion
-    # creates a circular dependency (because it's not possible to create the
-    # EB load balancer separately from the EB application itself).
-    environment_variables += [
-        ("APP_CLOUDFRONT_ID", If(
-            app_uses_cloudfront_condition,
-            Ref(app_distribution),
-            "",
-        )),
-        ("APP_CLOUDFRONT_DOMAINNAME", If(
-            app_uses_cloudfront_condition,
-            GetAtt(app_distribution, "DomainName"),
-            "",
-        )),
-    ]
 
 if es_domain:
     # not supported by GovCloud, so add it only if it was created
